Renamed exportActualMessage to exportEMLMessage.
Moved some functions out of index.ts into a new file.

Continued work on accepting command-line arguments to make code more user-friendly.
