+++ package.json
@@ -1,19 +1,23 @@
 {
-  "name": "coreapi-export-template",
+  "name": "front-exporter",
   "version": "1.0.0",
-  "description": "Export template application for Front Core API",
-  "license": "ISC",
+  "description": "Export application for Front Core API",
+  "license": "MIT",
   "scripts": {
     "start": "ts-node ./src/index.ts"
   },
   "dependencies": {
     "@types/needle": "^2.5.3",
     "@types/node": "^18.11.9",
+    "colors": "^1.4.0",
     "dotenv": "^16.0.3",
     "needle": "^3.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4",
     "winston": "^3.13.0",
     "yargs": "^17.7.2"
+  },
+  "devDependencies": {
+    "@types/yargs": "^17.0.32"
   }
 }
