+++ src/main.ts
@@ -1,0 +1,125 @@
+import { FrontExport, ExportOptions } from "./export";
+var colors = require('colors');
+import * as fs from 'fs';
+
+import { Logger } from "./logging";
+export const log = Logger.getLogger("M");
+
+// TODO: Move to .env
+// Set required options for the export
+const options : ExportOptions = {
+    shouldIncludeMessages: true,
+    exportAsEML: true, // If this option is set to true, the messages will only be exported as.eml files
+    shouldIncludeAttachments: true,
+    shouldIncludeComments: false
+}
+
+// List all inboxes available to the API key
+export function listInboxes() {
+    FrontExport.listInboxes()
+        .then(inboxes => {
+            log.info(`Listing Inboxes...`);
+            console.log(colors.yellow.underline("ID"), "\t\t", colors.blue.underline("Name"));
+            for (const inbox of inboxes) {
+                console.log(colors.yellow(inbox.id), "\t", colors.blue(inbox.name));
+            }
+        });
+}
+
+// TODO: Finish this function
+// Resume the export from where it left off
+export function resumeExport() {
+
+// load 2 json files called "allconvos.json" and "done.json" and compare them. if an entry in 
+// done.json matches allconvos.json, skip it. if it doesn't, add it to a new json file called 
+// "required.json".
+/*
+import path from "path";
+
+function getSubdirs(dir: string): string[] {
+    return fs.readdirSync(dir).filter((file: any) => fs.statSync(path.join(dir, file)).isDirectory());
+}
+
+function getCurrentProgress() {
+    const subdirs = getSubdirs("export");
+    fs.writeFileSync("done.json", JSON.stringify(subdirs));
+
+    const required = JSON.parse(fs.readFileSync("allconvos.json", "utf8"));
+    const subdirectories = JSON.parse(fs.readFileSync("done.json", "utf8"));
+    const missing = subdirectories.filter((subdir: string) => !required.includes(subdir));
+    fs.writeFileSync("required.json", JSON.stringify(missing));
+}
+*/
+
+
+
+    // Load "required.json" and add contents to a new array named requiredConversations
+    const requiredConversations: string[] = JSON.parse(fs.readFileSync('required.json', 'utf8'));
+    log.info(`Number Required: ${requiredConversations.length}`);
+
+
+    // Export specific conversations from a specific inbox, for example, the inbox with ID 'inb_ndb'
+    FrontExport.listInboxes()
+        .then(inboxes => {
+            log.info(`Starting export...`);
+            const inboxToExport = inboxes.find(inbox => inbox.id === 'inb_ndb'); // Export from a specific Inbox
+            if (inboxToExport) {
+                return FrontExport.exportSpecificConversations(requiredConversations, inboxToExport, options)
+                    .then(conversations => {
+                        console.log("Total:", conversations.length);
+                    });
+            } else {
+                throw new Error("Inbox with ID 'inb_ndb' not found.");
+            }
+        })
+        .catch(error => {
+            console.error("Error exporting conversations:", error);
+        });
+}
+
+
+// Export ALL conversations from ALL inboxes available to the API key
+export function exportAll() {
+    // Warning: May take a very long time to complete!
+    FrontExport.listInboxes()
+        .then(inboxes => {
+            for (const inbox of inboxes) {
+                FrontExport.exportInboxConversations(inbox, options)
+                .then(conversations => {
+                    log.info(`Total: ${conversations.length}`);
+                });
+            }
+        })
+        .catch(error => {
+            log.error("Error exporting conversations:", error);
+        });
+}
+
+
+// Export all conversations from a specific inbox, for example, an inbox with ID 'inb_abc'
+export function exportFromInbox(inboxID: string) {
+    FrontExport.listInboxes()
+    .then(inboxes => {
+        const inboxToExport = inboxes.find(inbox => inbox.id === inboxID); // Export from a specific Inbox
+        if (inboxToExport) {
+            return FrontExport.exportInboxConversations(inboxToExport, options)
+            .then(conversations => {
+                log.info(`Total: ${conversations.length}`);
+            });
+        } else {
+            throw new Error(`Inbox with ID ${inboxID} not found.`);
+        }
+    })
+    .catch(error => {
+        log.error("Error exporting conversations:", error);
+    });
+}
+
+// TODO: Fix this up so it accepts command line arguments, and converts date to unix timestamp
+// Export specific conversations matching the requiredConversations array
+export function exportSearch(searchArgs: string) {
+    FrontExport.exportSearchConversations('"google"', { after: 1704067200 }, ['open'], options)
+        .then(conversations => {
+            log.info(`Total: ${conversations.length}`);
+        })
+}
